Revert "chore : 패키지 구조 변경"
This reverts commit b665b49f66ea8bff90ea38c473f8fa93ea9de29a.
