+++ src/main/java/com/programmers/springbooturlshortener/domain/UrlRepository.java
@@ -1,0 +1,10 @@
+package com.programmers.springbooturlshortener.domain;
+
+import java.util.Optional;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UrlRepository extends JpaRepository<Url, Long> {
+
+	Optional<Url> findByOriginUrl(String originUrl);
+}
