+++ src/main/java/com/programmers/springbooturlshortener/domain/UrlResponseDto.java
@@ -1,4 +1,4 @@
-package com.programmers.springbooturlshortener.domain.url.dto;
+package com.programmers.springbooturlshortener.domain;
 
 public record UrlResponseDto(String originUrl, String shortUrl, Long requestCount) {
 }
