--- src/main/java/com/programmers/springbooturlshortener/domain/url/UrlRepository.java
@@ -1,10 +1,0 @@
-package com.programmers.springbooturlshortener.domain.url;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UrlRepository extends JpaRepository<Url, Long> {
-
-    Optional<Url> findByOriginUrl(String originUrl);
-}
